[dev.cc] runtime: change set_sec to take int64
Fixes build.
Tested that all these systems can make.bash.

TBR=austin
CC=golang-codereviews
https://golang.org/cl/177770043

+++ src/runtime/defs_freebsd_386.go
@@ -185,8 +185,8 @@ type timespec struct {
 	tv_nsec int32
 }
 
-func (ts *timespec) set_sec(x int32) {
-	ts.tv_sec = x
+func (ts *timespec) set_sec(x int64) {
+	ts.tv_sec = int32(x)
 }
 
 type timeval struct {
